+++ src/view_wrapper.tsx
@@ -1,8 +1,9 @@
 import * as PropTypes from 'prop-types';
 import { equals, keys, merge, mergeAll, omit, pick } from 'ramda';
 import * as React from 'react';
-import { Container, DelegateDef, Environment, MessageConstructor } from './app';
+import { Container, DelegateDef, MessageConstructor } from './app';
 import ErrorComponent from './components/error';
+import { Environment } from './environment';
 import ExecContext from './exec_context';
 import { Activate, Deactivate, Refresh } from './message';
